+++ fastapi_limiter/__init__.py
@@ -26,7 +26,6 @@ async def default_callback(request: Request, response: Response, pexpire: int):
     :return:
     """
     expire = ceil(pexpire / 1000)
-
     raise HTTPException(
         HTTP_429_TOO_MANY_REQUESTS, "Too Many Requests", headers={"Retry-After": str(expire)}
     )
@@ -35,11 +34,28 @@ async def default_callback(request: Request, response: Response, pexpire: int):
 class FastAPILimiter:
     redis: aioredis.Redis = None
     prefix: str = None
+    lua_sha: str = None
     identifier: Callable = None
     callback: Callable = None
+    lua_script = """local key = KEYS[1]
+local limit = tonumber(ARGV[1])
+local expire_time = ARGV[2]
+
+local current = tonumber(redis.call('get', key) or "0")
+if current > 0 then
+ if current + 1 > limit then
+ return redis.call("PTTL",key)
+ else
+        redis.call("INCR", key)
+ return 0
+ end
+else
+    redis.call("SET", key, 1,"px",expire_time)
+ return 0
+end"""
 
     @classmethod
-    def init(
+    async def init(
         cls,
         redis: aioredis.Redis,
         prefix: str = "fastapi-limiter",
@@ -50,3 +66,9 @@ class FastAPILimiter:
         cls.prefix = prefix
         cls.identifier = identifier
         cls.callback = callback
+        cls.lua_sha = await redis.script_load(cls.lua_script)
+
+    @classmethod
+    async def close(cls):
+        cls.redis.close()
+        await cls.redis.wait_closed()