+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## 0.1
 
+### 0.1.4
+
+- Now use `lua` script.
+- **Break change**: You shoud call `FastAPILimiter.init` with `async`.
+
+```python
+    await FastAPILimiter.init(redis)
+```
+
 ### 0.1.3
 
 - Support multiple rate strategy for one route. (#3)